# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AAELink HTTP API",
    "version": "0.0.41-alpha",
    "summary": "Auto-generated route inventory. Method/path surface only — request/response schemas not yet introspected."
  },
  "servers": [
    {
      "url": "https://example.com",
      "description": "Replace with your AAELink deployment URL"
    }
  ],
...
  "paths": {
...
    "/api/drafts": {
      "get": {
        "summary": "GET /api/drafts",
        "tags": [
          "drafts"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "PUT /api/drafts",
        "tags": [
          "drafts"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "DELETE /api/drafts",
        "tags": [
          "drafts"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/lists": {
      "get": {
        "summary": "GET /api/lists",
        "tags": [
          "lists"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "POST /api/lists",
        "tags": [
          "lists"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}